refactor: entities/auth API 레이어 분리 및 mutation 제거
