+++ src/entities/auth/index.ts
@@ -1,9 +1,2 @@
 export type { AuthResponse, AuthStore, AuthUser } from './model';
-export {
-  authApi,
-  useAuthStore,
-  useLoginMutation,
-  useLogoutMutation,
-  useSignUpMutation,
-  useSocialLoginMutation,
-} from './model';
+export { authApi, useAuthStore } from './model';
