+++ src/entities/auth/model/index.ts
@@ -1,11 +1,5 @@
-export type { AuthResponse } from './auth.api';
-export { authApi } from './auth.api';
-export {
-  useLoginMutation,
-  useLogoutMutation,
-  useSignUpMutation,
-  useSocialLoginMutation,
-} from './auth.queries';
+export type { AuthResponse } from '../api/auth.api';
+export { authApi } from '../api/auth.api';
 export type { AuthUser } from './auth.types';
 export type { AuthStore } from './auth-store';
 export { useAuthStore } from './auth-store';
